+++ Tests/Runtime/CleanBundleCacheTests.cs
@@ -187,8 +187,8 @@ namespace AddressableAssetsIntegrationTests
 
 #endif
 
+#if !UNITY_PS5
         [UnityTest]
-        [Platform(Exclude = "PS5")]
         public IEnumerator WhenValidCatalogId_RemovesNonReferencedBundlesFromCache([Values(true, false)] bool forceSingleThreading)
         {
 #if ENABLE_CACHING
@@ -218,9 +218,10 @@ namespace AddressableAssetsIntegrationTests
             yield return null;
 #endif
         }
+#endif
 
+#if !UNITY_PS5
         [UnityTest]
-        [Platform(Exclude = "PS5")]
         public IEnumerator WhenCatalogIdListNull_UsesLoadedCatalogs_AndRemovesNonReferencedBundlesFromCache()
         {
 #if ENABLE_CACHING
@@ -249,6 +250,7 @@ namespace AddressableAssetsIntegrationTests
             yield return null;
 #endif
         }
+#endif
 
         [UnityTest]
         public IEnumerator WhenCatalogIdListNull_AndUsingFastMode_ReturnsException()